Remove repository
DbContext already acts as a repository. If needed,
just add methods to it or to DbSet to simplify its use.
