--- src/ApplicationCore/Interfaces/IAsyncRepository.cs
@@ -1,18 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Threading.Tasks;
-using RecipeManager.ApplicationCore.Entities;
-
-namespace RecipeManager.ApplicationCore.Interfaces
-{
-    public interface IAsyncRepository<T> where T : BaseEntity
-    {
-        Task<T> GetByIdAsync(Guid id);
-
-        Task<IReadOnlyList<T>> ListAllAsync();
-
-        Task<IReadOnlyList<T>> ListAsync(ISpecification<T> spec);
-
-        Task<int> CountAsync(ISpecification<T> spec);
-    }
-}
