--- src/ApplicationCore/Interfaces/IRecipeRepository.cs
@@ -1,8 +1,0 @@
-﻿namespace RecipeManager.ApplicationCore.Interfaces
-{
-    using RecipeManager.ApplicationCore.Entities;
-
-    public interface IRecipeRepository : IAsyncRepository<Recipe>
-    {
-    }
-}
