--- src/Infrastructure/Data/RecipeRepository.cs
@@ -1,13 +1,0 @@
-﻿namespace RecipeManager.Infrastructure.Data
-{
-    using RecipeManager.ApplicationCore.Entities;
-    using RecipeManager.ApplicationCore.Interfaces;
-
-    public class RecipeRepository : EfRepository<Recipe>, IRecipeRepository
-    {
-        public RecipeRepository(RecipeApiDbContext context)
-            : base(context)
-        {
-        }
-    }
-}
